Fix: requests get duplicated using shared_prefix datagen when multi-turn chat disabled  (#293)
### Background
A bug was causing requests to be duplicated when using the shared_prefix
with multi-turn chat disabled. This happened because the load generator
was creating a standalone request queue for each worker, and then
broadcasting each incoming request to all worker queues.

### Fix
The fix ensures that this standalone queue feature is disabled when
multi-turn chat is not active, preventing the duplication of requests.

### Credit
Credit for reporting this bug goes to @diamondburned